working...

+++ ServerController.js
@@ -22,32 +22,39 @@ class Controller{
         })
     }
     handle_subscribe(predicate, args, ticket){
+        console.log('subscribe', predicate, args, ticket)
         args = args || []
         let ret = {ticket: ticket, type: 'subscribe'}
         let pred = this['subs_' + predicate](...args)
-        pred.changes({includeInitial: true}).run(this.conn, (err, cursor)=>{
+        pred.changes({includeInitial: true, includeStates: true}).run(this.conn, (err, cursor)=>{
             this.cursors.push(cursor)
-            cursor.each((err, data)=>{
-                let type
-                if(!data.old_val){
-                    type = 'add'
-                    data.newVal = data.new_val
-                    delete data.old_val
-                    delete data.new_val
-                }else if(!data.new_val){
-                    type = 'delete'
-                    delete data.old_val
-                    delete data.new_val
-                }else{
-                    type = 'update'
-                    data.newVal = data.new_val
-                    delete data.old_val
-                    delete data.new_val
+            cursor.each((err, data)=> {
+                    if (data.state) {
+                        data = {type: data.state, ticket: ticket}
+                        this.ws.send(JSON.stringify(data))
+                    } else {
+                        let type
+                        if (!data.old_val) {
+                            type = 'add'
+                            data.newVal = data.new_val
+                            delete data.old_val
+                            delete data.new_val
+                        } else if (!data.new_val) {
+                            type = 'delete'
+                            delete data.old_val
+                            delete data.new_val
+                        } else {
+                            type = 'update'
+                            data.newVal = data.new_val
+                            delete data.old_val
+                            delete data.new_val
+                        }
+                        ret.type = type
+                        ret.data = data
+                        ret.predicate = predicate
+                        this.ws.send(JSON.stringify(ret))
+                    }
                 }
-                ret.type = type
-                ret.data=data
-                ret.predicate = predicate
-                this.ws.send(JSON.stringify(ret))}
             )
             // cursor.on('data', (change) => {ret.data=change; this.ws.send(JSON.stringify(ret))})
         })
